extract mutation tracker
